Separating out components of transformer into their own files
