+++ parlai/agents/transformer/functions.py
@@ -1,0 +1,28 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Pure functions used in transformer implementation.
+"""
+
+import numpy as np
+import torch
+
+
+def create_position_codes(n_pos, dim, out):
+    """
+    Create positional codes and store them in ``out``.
+    """
+    position_enc = np.array(
+        [
+            [pos / np.power(10000, 2 * j / dim) for j in range(dim // 2)]
+            for pos in range(n_pos)
+        ]
+    )
+
+    out.detach_()
+    out.requires_grad = False
+    out[:, 0::2] = torch.FloatTensor(np.sin(position_enc)).type_as(out)
+    out[:, 1::2] = torch.FloatTensor(np.cos(position_enc)).type_as(out)
