+++ parlai/agents/transformer/modules/decoder.py
@@ -1,0 +1,410 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Transformer decoder implementations.
+"""
+
+from typing import Dict, Tuple, Optional
+
+import numpy as np
+import torch
+import torch.nn as nn
+
+from parlai.agents.transformer.functions import create_position_codes
+from parlai.agents.transformer.modules.attention import MultiHeadAttention
+from parlai.agents.transformer.modules.ffn import TransformerFFN
+from parlai.agents.transformer.modules.layer_norm import LayerNorm, normalize
+from parlai.utils.misc import warn_once
+from parlai.utils.torch import PipelineHelper
+
+
+class TransformerDecoder(nn.Module):
+    """
+    Transformer Decoder layer.
+
+    :param int n_heads: the number of multihead attention heads.
+    :param int n_layers: number of transformer layers.
+    :param int embedding_size: the embedding sizes. Must be a multiple of n_heads.
+    :param int ffn_size: the size of the hidden layer in the FFN
+    :param embedding: an embedding matrix for the bottom layer of the transformer.
+        If none, one is created for this encoder.
+    :param float dropout: Dropout used around embeddings and before layer
+        layer normalizations. This is used in Vaswani 2017 and works well on
+        large datasets.
+    :param float attention_dropout: Dropout performed after the multhead attention
+        softmax. This is not used in Vaswani 2017.
+    :param float relu_attention: Dropout used after the ReLU in the FFN. Not used
+        in Vaswani 2017, but used in Tensor2Tensor.
+    :param int padding_idx: Reserved padding index in the embeddings matrix.
+    :param bool learn_positional_embeddings: If off, sinusoidal embeddings are
+        used. If on, position embeddings are learned from scratch.
+    :param bool embeddings_scale: Scale embeddings relative to their dimensionality.
+        Found useful in fairseq.
+    :param int n_positions: Size of the position embeddings matrix.
+    """
+
+    def __init__(
+        self,
+        n_heads,
+        n_layers,
+        embedding_size,
+        ffn_size,
+        vocabulary_size,
+        embedding=None,
+        dropout=0.0,
+        attention_dropout=0.0,
+        relu_dropout=0.0,
+        embeddings_scale=True,
+        learn_positional_embeddings=False,
+        padding_idx=None,
+        n_positions=1024,
+        n_segments=0,
+        variant='aiayn',
+        activation='relu',
+    ):
+        super().__init__()
+        self.embedding_size = embedding_size
+        self.ffn_size = ffn_size
+        self.n_layers = n_layers
+        self.n_heads = n_heads
+        self.dim = embedding_size
+        self.activation = activation
+        self.variant = variant
+
+        self.embeddings_scale = embeddings_scale
+        self.dropout = nn.Dropout(p=dropout)  # --dropout
+
+        self.n_positions = n_positions
+        self.out_dim = embedding_size
+        assert (
+            embedding_size % n_heads == 0
+        ), 'Transformer embedding size must be a multiple of n_heads'
+
+        self.embeddings = embedding
+
+        if (
+            self.variant == 'xlm'
+            or self.variant == 'prelayernorm'
+            or self.variant == 'bart'
+        ):
+            self.norm_embeddings = LayerNorm(self.dim)
+            if self.variant == 'xlm':
+                warn_once(
+                    'DEPRECATED: XLM should only be used for backwards compatibility, '
+                    'as it involves a less-stable layernorm operation.'
+                )
+        elif self.variant == 'aiayn':
+            pass
+        else:
+            raise ValueError("Can't handle --variant {}".format(self.variant))
+
+        # create the positional embeddings
+        self.position_embeddings = nn.Embedding(n_positions, embedding_size)
+        if not learn_positional_embeddings:
+            create_position_codes(
+                n_positions, embedding_size, out=self.position_embeddings.weight
+            )
+        else:
+            nn.init.normal_(self.position_embeddings.weight, 0, embedding_size ** -0.5)
+
+        # build the model
+        self.layers = nn.ModuleList()
+        for _ in range(self.n_layers):
+            self.layers.append(
+                TransformerDecoderLayer(
+                    n_heads,
+                    embedding_size,
+                    ffn_size,
+                    attention_dropout=attention_dropout,
+                    relu_dropout=relu_dropout,
+                    dropout=dropout,
+                    activation=activation,
+                    variant=variant,
+                )
+            )
+
+    def forward_embedding(
+        self,
+        input: torch.LongTensor,
+        positions: Optional[torch.LongTensor] = None,
+        segments: Optional[torch.LongTensor] = None,
+    ):
+        """
+        Embed tokens prior to feeding into transformer.
+
+        :param LongTensor[batch, seqlen] input:
+            The target input IDs
+        :param LongTensor[batch, seqlen] positions:
+            Positions for input IDs. If None, computes defaults.
+        :param LongTensor[batch, seqlen] segements:
+            Segment IDs for extra embedding features. If None, not used.
+
+        :return (tensor, mask):
+            embeded input and mask
+        """
+        tensor = self.embeddings(input)
+        if self.embeddings_scale:
+            tensor = tensor * np.sqrt(self.dim)
+        if self.variant == 'xlm':
+            tensor = normalize(tensor, self.norm_embeddings)
+        if positions.max().item() > self.n_positions:
+            warn_once(
+                'You are inputting a sequence of {x} length, but only have '
+                '--n-positions {y}. Set --truncate or increase --n-positions'.format(
+                    x=positions.max().item(), y=self.n_positions
+                )
+            )
+        tensor = tensor + self.position_embeddings(positions).expand_as(tensor)
+        if self.variant == 'bart':
+            tensor = normalize(tensor, self.norm_embeddings)
+
+        return tensor
+
+    def forward_layers(
+        self,
+        tensor: torch.Tensor,
+        encoder_output: torch.Tensor,
+        encoder_mask: torch.Tensor,
+        incr_state: Dict[int, torch.Tensor],
+    ) -> Tuple[torch.Tensor, Dict[str, torch.Tensor]]:
+        """
+        Forward pass of decoder layers.
+
+        :param tensor:
+            embedded input tensor for the decoder
+        :param enc_out:
+            encoder outputs
+        :param enc_mask:
+            encoder output mask
+        :param incr_state:
+            Dict mapping layer_idx to incremental state
+
+        :return (tensor, new_incr_state):
+            return encoding after applying decoder layers, as well
+            as new incremental decoding state.
+        """
+        new_incr_state = {}
+        if getattr(self.layers, 'is_model_parallel', False):
+            tensor, new_incr_state = self._apply_model_parallel(
+                tensor, encoder_output, encoder_mask, incr_state
+            )
+        else:
+            for idx, layer in enumerate(self.layers):
+                tensor, new_incr_state[idx] = layer(
+                    x=tensor,
+                    encoder_output=encoder_output,
+                    encoder_mask=encoder_mask,
+                    incr_state=incr_state.get(idx),
+                )
+
+        return tensor, new_incr_state
+
+    def forward(self, input, encoder_state, incr_state=None):
+        """
+        Forward pass.
+
+        :param LongTensor[batch,seqlen] input:
+            The decoder inputs (partial or full decoded token IDs).
+        :param encoder_state:
+            Output from the encoder module forward pass.
+        :param incr_state:
+            The incremental state: a dictionary whose keys index the layers and whose
+            values contain the incremental state for each layer.
+        """
+        encoder_output, encoder_mask = encoder_state
+
+        seq_len = input.size(1)
+        positions = input.new(seq_len).long()
+        positions = torch.arange(seq_len, out=positions).unsqueeze(0)
+
+        if incr_state is not None:
+            # We're doing incremental decoding, so select only the most recent position
+            input = input[:, -1:]
+            if positions is not None:
+                positions = positions[:, -1:]
+        else:
+            incr_state = {}
+
+        tensor = self.forward_embedding(input, positions)
+
+        tensor = self.dropout(tensor)  # --dropout
+
+        tensor, new_incr_state = self.forward_layers(
+            tensor, encoder_output, encoder_mask, incr_state
+        )
+
+        if self.variant == 'prelayernorm':
+            tensor = normalize(tensor, self.norm_embeddings)
+
+        return tensor, new_incr_state
+
+    def _apply_model_parallel(self, tensor, encoder_output, encoder_mask, incr_state):
+        """
+        Pipeline application of model parallelism.
+        """
+        chunks = PipelineHelper.split(
+            (tensor, encoder_output, encoder_mask, incr_state)
+        )
+        work_items = PipelineHelper.schedule_work_items(self.layers, chunks)
+
+        new_incr_state = {i: [] for i, _ in enumerate(self.layers)}
+
+        for chunk_idx, layer_nos, next_device in work_items:
+            s_tensor, s_enc_out, s_enc_mask, s_incr_state = chunks[chunk_idx]
+            for layer_no in layer_nos:
+                s_tensor, nis = self.layers[layer_no](
+                    x=s_tensor,
+                    encoder_output=s_enc_out,
+                    encoder_mask=s_enc_mask,
+                    incr_state=s_incr_state.get(layer_no),
+                )
+                new_incr_state[layer_no].append(nis)
+            # don't move incr state, it's always on the correct device
+            s_tensor, s_enc_out, s_enc_mask = PipelineHelper.chunk_to(
+                (s_tensor, s_enc_out, s_enc_mask), next_device
+            )
+            chunks[chunk_idx] = (s_tensor, s_enc_out, s_enc_mask, s_incr_state)
+
+        tensor_out = PipelineHelper.join([c[0] for c in chunks])
+        new_incr_state = {
+            layer_no: PipelineHelper.join(pieces)
+            for layer_no, pieces in new_incr_state.items()
+        }
+
+        return tensor_out, new_incr_state
+
+
+class TransformerDecoderLayer(nn.Module):
+    """
+    Implements a single Transformer decoder layer.
+
+    Decoder layers are similar to encoder layers but:
+
+    1. Self-attention is limited in a casaul (auto-regressive) manner.
+    2. Attend over all of the encoder states.
+    """
+
+    def __init__(
+        self,
+        n_heads,
+        embedding_size,
+        ffn_size,
+        attention_dropout=0.0,
+        relu_dropout=0.0,
+        dropout=0.0,
+        activation='relu',
+        variant='aiayn',
+    ):
+        super().__init__()
+        self.dim = embedding_size
+        self.ffn_dim = ffn_size
+        self.variant = variant
+        self.activation = activation
+        self.dropout = nn.Dropout(p=dropout)
+
+        self.self_attention = MultiHeadAttention(
+            n_heads, embedding_size, dropout=attention_dropout
+        )
+        self.norm1 = LayerNorm(embedding_size)
+
+        self.encoder_attention = MultiHeadAttention(
+            n_heads, embedding_size, dropout=attention_dropout
+        )
+        self.norm2 = LayerNorm(embedding_size)
+
+        self.ffn = TransformerFFN(
+            embedding_size, ffn_size, relu_dropout=relu_dropout, activation=activation
+        )
+        self.norm3 = LayerNorm(embedding_size)
+
+    def forward(self, x, encoder_output, encoder_mask, incr_state=None):
+        """
+        Forward pass.
+
+        The incremental state is a dict with values for self- and encoder-attention
+        states.
+        """
+
+        if incr_state is None:
+            incr_state = {}
+
+        decoder_mask = self._create_selfattn_mask(x)
+        # first self attn
+        residual = x
+        if self.variant == 'prelayernorm':
+            x = normalize(x, self.norm1)
+
+        # don't peak into the future!
+        x, final_self_attn_incr_state = self.self_attention(
+            query=x,
+            mask=decoder_mask,
+            incr_state=incr_state.get('self_attn'),
+            static_kv=False,
+        )[:2]
+        x = self.dropout(x)  # --dropout
+        x = x + residual
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            x = normalize(x, self.norm1)
+
+        residual = x
+        # encoder_attn_layer_norm norm 2
+        if self.variant == 'prelayernorm':
+            x = normalize(x, self.norm2)
+        x, final_encoder_attn_incr_state = self.encoder_attention(
+            query=x,
+            key=encoder_output,
+            value=encoder_output,
+            mask=encoder_mask,
+            incr_state=incr_state.get('encoder_attn'),
+            static_kv=True,
+        )[:2]
+        x = self.dropout(x)  # --dropout
+        x = residual + x
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            x = normalize(x, self.norm2)
+
+        # finally the ffn
+        residual = x
+        if self.variant == 'prelayernorm':
+            x = normalize(x, self.norm3)
+        x = self.ffn(x)
+        x = self.dropout(x)  # --dropout
+        x = residual + x
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            x = normalize(x, self.norm3)
+
+        new_incr_state = {
+            'self_attn': final_self_attn_incr_state,
+            'encoder_attn': final_encoder_attn_incr_state,
+        }
+        return x, new_incr_state
+
+    def _create_selfattn_mask(self, x):
+        # figure out how many timestamps we need
+        bsz = x.size(0)
+        time = x.size(1)
+        # make sure that we don't look into the future
+        mask = torch.tril(x.new(time, time).fill_(1))
+        # broadcast across batch
+        mask = mask.unsqueeze(0).expand(bsz, -1, -1)
+        return mask
+
+    def reorder_incremental_state(
+        self, incremental_state: Dict[str, dict], inds: torch.Tensor
+    ) -> Dict[str, dict]:
+        """
+        Reorder all incremental-state tensors for this layer.
+        """
+        attn_types = {
+            'self_attn': self.self_attention,
+            'encoder_attn': self.encoder_attention,
+        }
+        return {
+            attn_type: attn.reorder_incremental_state(
+                incremental_state[attn_type], inds
+            )
+            for attn_type, attn in attn_types.items()
+        }
