+++ parlai/agents/transformer/modules/encoder.py
@@ -1,0 +1,369 @@
+#!/usr/bin/env python3
+# Copyright (c) Facebook, Inc. and its affiliates.
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Transformer encoder implementations.
+"""
+
+from typing import Tuple, Optional, Union
+
+import numpy as np
+import torch
+import torch.nn as nn
+
+from parlai.agents.transformer.functions import create_position_codes
+from parlai.agents.transformer.modules.attention import MultiHeadAttention
+from parlai.agents.transformer.modules.ffn import TransformerFFN
+from parlai.agents.transformer.modules.layer_norm import LayerNorm, normalize
+from parlai.utils.misc import warn_once
+from parlai.utils.torch import PipelineHelper
+
+
+class TransformerEncoder(nn.Module):
+    """
+    Transformer encoder module.
+
+    :param int n_heads: the number of multihead attention heads.
+    :param int n_layers: number of transformer layers.
+    :param int embedding_size: the embedding sizes. Must be a multiple of n_heads.
+    :param int ffn_size: the size of the hidden layer in the FFN
+    :param embedding: an embedding matrix for the bottom layer of the transformer.
+        If none, one is created for this encoder.
+    :param float dropout: Dropout used around embeddings and before layer
+        layer normalizations. This is used in Vaswani 2017 and works well on
+        large datasets.
+    :param float attention_dropout: Dropout performed after the multhead attention
+        softmax. This is not used in Vaswani 2017.
+    :param float relu_attention: Dropout used after the ReLU in the FFN. Not used
+        in Vaswani 2017, but used in Tensor2Tensor.
+    :param int padding_idx: Reserved padding index in the embeddings matrix.
+    :param bool learn_positional_embeddings: If off, sinusoidal embeddings are
+        used. If on, position embeddings are learned from scratch.
+    :param bool embeddings_scale: Scale embeddings relative to their dimensionality.
+        Found useful in fairseq.
+    :param bool reduction: If true, returns the mean vector for the entire encoding
+        sequence.
+    :param int n_positions:
+        Size of the position embeddings matrix.
+    :param int n_segments:
+        Number of segments/lang/sentence embeddings.
+    :param activation:
+        Type of nonlinear activation. Can be relu or gelu.
+    :param variant:
+        Which transformer architecture to use. Could be AIAYN or XLM.
+        Future versions may support things like GPT-2, ...
+    :param output_scaling:
+        Scale the outputs by a given scalar
+    """
+
+    def __init__(
+        self,
+        n_heads,
+        n_layers,
+        embedding_size,
+        ffn_size,
+        vocabulary_size,
+        embedding=None,
+        dropout=0.0,
+        attention_dropout=0.0,
+        relu_dropout=0.0,
+        padding_idx=0,
+        learn_positional_embeddings=False,
+        embeddings_scale=False,
+        reduction_type='mean',
+        n_positions=1024,
+        activation='relu',
+        variant='aiayn',
+        n_segments=0,
+        output_scaling=1.0,
+    ):
+        super(TransformerEncoder, self).__init__()
+
+        self.embedding_size = embedding_size
+        self.ffn_size = ffn_size
+        self.n_layers = n_layers
+        self.n_heads = n_heads
+        self.dim = embedding_size
+        self.embeddings_scale = embeddings_scale
+        self.reduction_type = reduction_type
+        self.padding_idx = padding_idx
+        # this is --dropout, not --relu-dropout or --attention-dropout
+        self.dropout_frac = dropout
+        self.dropout = nn.Dropout(p=self.dropout_frac)
+        self.variant = variant
+        self.n_segments = n_segments
+
+        self.n_positions = n_positions
+        self.out_dim = embedding_size
+        assert (
+            embedding_size % n_heads == 0
+        ), 'Transformer embedding size must be a multiple of n_heads'
+
+        # check input formats:
+        if embedding is not None:
+            assert (
+                embedding_size is None or embedding_size == embedding.weight.shape[1]
+            ), "Embedding dim must match the embedding size."
+
+        if embedding is not None:
+            self.embeddings = embedding
+        else:
+            raise AssertionError(
+                "This code should not execute. Left here in case we want to enable it."
+            )
+            assert padding_idx is not None
+            self.embeddings = nn.Embedding(
+                vocabulary_size, embedding_size, padding_idx=padding_idx
+            )
+            nn.init.normal_(self.embeddings.weight, 0, embedding_size ** -0.5)
+
+        # create the positional embeddings
+        self.position_embeddings = nn.Embedding(n_positions, embedding_size)
+        if not learn_positional_embeddings:
+            create_position_codes(
+                n_positions, embedding_size, out=self.position_embeddings.weight
+            )
+        else:
+            nn.init.normal_(self.position_embeddings.weight, 0, embedding_size ** -0.5)
+
+        # embedding normalization
+        if (
+            self.variant == 'xlm'
+            or self.variant == 'prelayernorm'
+            or self.variant == 'bart'
+        ):
+            self.norm_embeddings = LayerNorm(self.dim)
+        elif self.variant == 'aiayn':
+            pass
+        else:
+            raise ValueError("Can't handle --variant {}".format(self.variant))
+
+        if self.n_segments >= 1:
+            self.segment_embeddings = nn.Embedding(self.n_segments, self.dim)
+
+        # build the model
+        self.layers = nn.ModuleList()
+        for _ in range(self.n_layers):
+            self.layers.append(
+                TransformerEncoderLayer(
+                    n_heads,
+                    embedding_size,
+                    ffn_size,
+                    attention_dropout=attention_dropout,
+                    relu_dropout=relu_dropout,
+                    dropout=dropout,
+                    variant=variant,
+                    activation=activation,
+                )
+            )
+        self.output_scaling = output_scaling
+
+    def forward_embedding(
+        self,
+        input: torch.LongTensor,
+        positions: Optional[torch.LongTensor] = None,
+        segments: Optional[torch.LongTensor] = None,
+    ) -> Tuple[torch.Tensor, torch.BoolTensor]:
+        """
+        Embed tokens prior to feeding into transformer.
+
+        :param LongTensor[batch,seqlen] input:
+            The input IDs
+        :param LongTensor[batch,seqlen] positions:
+            Positions for input IDs
+        :param LongTensor[batch,seqlen]:
+            If provided, additionally adds ``segments`` as extra embedding features.
+
+        :return (tensor, mask):
+            return embedded input and mask
+        """
+        mask = input != self.padding_idx
+        if positions is None:
+            positions = (mask.cumsum(dim=1, dtype=torch.int64) - 1).clamp_(min=0)
+        tensor = self.embeddings(input)
+        if self.embeddings_scale:
+            tensor = tensor * np.sqrt(self.dim)
+
+        if positions.max().item() > self.n_positions:
+            warn_once(
+                'You are inputting a sequence of {x} length, but only have '
+                '--n-positions {y}. Set --truncate or increase --n-positions'.format(
+                    x=positions.max().item(), y=self.n_positions
+                )
+            )
+        position_embs = self.position_embeddings(positions).expand_as(tensor)
+        tensor = tensor + position_embs
+
+        if self.n_segments >= 1:
+            if segments is None:
+                segments = torch.zeros_like(input)  # type: ignore
+            tensor = tensor + self.segment_embeddings(segments)
+
+        return tensor, mask
+
+    def forward_layers(
+        self, tensor: torch.Tensor, mask: torch.BoolTensor
+    ) -> torch.Tensor:
+        """
+        Apply transformer layers to input.
+
+        :param tensor:
+            embedded input
+        :param mask:
+            mask of input
+
+        :return tensor:
+            return embedding after applying transformer layers
+        """
+        if getattr(self.layers, 'is_model_parallel', False):
+            # factored out for readability. It is equivalent to the other
+            # condition
+            tensor = self._apply_model_parallel(tensor, mask)
+        else:
+            for i in range(self.n_layers):
+                tensor = self.layers[i](tensor, mask)
+
+        return tensor
+
+    def reduce_output(
+        self, tensor: torch.Tensor, mask: torch.BoolTensor
+    ) -> Tuple[torch.Tensor, Optional[torch.BoolTensor]]:
+        """
+        Reduce transformer output at end of forward pass.
+
+        :param tensor:
+            encoded input
+        :param mask:
+            mask for encoded input
+
+        :return (tensor, mask):
+            returns the reduced tensor, and mask if appropriate
+        """
+        tensor *= self.output_scaling
+        if self.reduction_type == 'first':
+            return tensor[:, 0, :], None
+        elif self.reduction_type == 'max':
+            return tensor.max(dim=1)[0], None
+        elif self.reduction_type == 'mean':
+            divisor = mask.float().sum(dim=1).unsqueeze(-1).clamp(min=1).type_as(tensor)
+            output = tensor.sum(dim=1) / divisor
+            return output, None
+        elif self.reduction_type is None or 'none' in self.reduction_type:
+            return tensor, mask
+        else:
+            raise ValueError(
+                "Can't handle --reduction-type {}".format(self.reduction_type)
+            )
+
+    def forward(  # type: ignore
+        self,
+        input: torch.LongTensor,
+        positions: Optional[torch.LongTensor] = None,
+        segments: Optional[torch.LongTensor] = None,
+    ) -> Union[torch.Tensor, Tuple[torch.Tensor, torch.BoolTensor]]:
+        """
+        Forward pass.
+
+        :param LongTensor[batch,seqlen] input:
+            The input IDs
+        :param LongTensor[batch,seqlen] positions:
+            Positions for input IDs
+        :param LongTensor[batch,seqlen] segments:
+            If provided, additionally adds ``segments`` as extra embedding features.
+        """
+        # embed input
+        tensor, mask = self.forward_embedding(input, positions, segments)
+
+        if self.variant == 'xlm' or self.variant == 'bart':
+            tensor = normalize(tensor, self.norm_embeddings)
+
+        # --dropout on the embeddings
+        tensor = self.dropout(tensor)
+
+        tensor *= mask.unsqueeze(-1).type_as(tensor)
+
+        # apply transformer layers
+        tensor = self.forward_layers(tensor, mask)
+
+        if self.variant == 'prelayernorm':
+            tensor = normalize(tensor, self.norm_embeddings)
+
+        # reduce output
+        tensor, out_mask = self.reduce_output(tensor, mask)
+        if out_mask is not None:
+            return tensor, out_mask
+        else:
+            return tensor
+
+    def _apply_model_parallel(self, tensor, mask):
+        """
+        Pipeline application of model parallelism.
+        """
+        chunks = PipelineHelper.split((tensor, mask))
+        work_items = PipelineHelper.schedule_work_items(self.layers, chunks)
+
+        for chunk_idx, layer_nos, next_device in work_items:
+            s_tensor, s_mask = chunks[chunk_idx]
+            for layer_no in layer_nos:
+                s_tensor = self.layers[layer_no](s_tensor, s_mask)
+            chunks[chunk_idx] = PipelineHelper.chunk_to((s_tensor, s_mask), next_device)
+
+        tensor_out, mask_out = PipelineHelper.join(chunks)
+        return tensor_out
+
+
+class TransformerEncoderLayer(nn.Module):
+    """
+    Implements a single Transformer encoder layer.
+    """
+
+    def __init__(
+        self,
+        n_heads,
+        embedding_size,
+        ffn_size,
+        attention_dropout=0.0,
+        relu_dropout=0.0,
+        dropout=0.0,
+        activation='relu',
+        variant=None,
+    ):
+        super().__init__()
+        self.dim = embedding_size
+        self.ffn_dim = ffn_size
+        self.activation = activation
+        self.variant = variant
+        self.attention = MultiHeadAttention(
+            n_heads, embedding_size, dropout=attention_dropout  # --attention-dropout
+        )
+        self.norm1 = LayerNorm(embedding_size)
+        self.ffn = TransformerFFN(
+            embedding_size,
+            ffn_size,
+            relu_dropout=relu_dropout,
+            activation=self.activation,
+        )
+        self.norm2 = LayerNorm(embedding_size)
+        self.dropout = nn.Dropout(p=dropout)
+
+    def forward(self, tensor, mask):
+        """
+        Forward pass.
+        """
+        residual = tensor
+        if self.variant == 'prelayernorm':
+            tensor = normalize(tensor, self.norm1)
+        attended_tensor = self.attention(tensor, mask=mask)[0]
+        tensor = residual + self.dropout(attended_tensor)
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            tensor = normalize(tensor, self.norm1)
+        residual = tensor
+        if self.variant == 'prelayernorm':
+            tensor = normalize(tensor, self.norm2)
+        tensor = residual + self.dropout(self.ffn(tensor))
+        if self.variant == 'aiayn' or self.variant == 'xlm' or self.variant == 'bart':
+            tensor = normalize(tensor, self.norm2)
+        tensor *= mask.unsqueeze(-1).type_as(tensor)
+        return tensor
